+++ CHANGELOG.md
@@ -14,6 +14,7 @@ and what APIs have changed, if applicable.
 
 ## [Unreleased]
 - Add fluent client api for subresources
+- Update fluent client APIs to include projection mask as input parameter.
 
 ## [29.17.0] - 2021-03-23
 - Implement D2 cluster subsetting.